Move chain to provider, remove notif sink
