+++ crates/consensus/auto-seal/src/task.rs
@@ -1,13 +1,13 @@
 use crate::{mode::MiningMode, Storage};
 use futures_util::{future::BoxFuture, FutureExt, StreamExt};
 use reth_beacon_consensus::BeaconEngineMessage;
-use reth_interfaces::{consensus::ForkchoiceState};
+use reth_interfaces::consensus::ForkchoiceState;
 use reth_primitives::{
     constants::{EMPTY_RECEIPTS, EMPTY_TRANSACTIONS},
     proofs, Block, BlockBody, ChainSpec, Header, IntoRecoveredTransaction, ReceiptWithBloom,
-    EMPTY_OMMER_ROOT, U256,
+    SealedBlockWithSenders, EMPTY_OMMER_ROOT, U256,
 };
-use reth_provider::{StateProviderFactory, CanonStateNotificationSender};
+use reth_provider::{CanonStateNotificationSender, Chain, StateProviderFactory};
 use reth_revm::{
     database::{State, SubState},
     executor::Executor,
@@ -165,24 +165,30 @@ where
 
                     trace!(target: "consensus::auto", transactions=?&block.body, "executing transactions");
 
-                    match executor.execute_transactions(&block, U256::ZERO, None) {
-                        Ok((res, gas_used)) => {
+                    let senders = block
+                        .body
+                        .iter()
+                        .map(|tx| tx.recover_signer())
+                        .collect::<Option<Vec<_>>>()?;
+
+                    match executor.execute_transactions(&block, U256::ZERO, Some(senders.clone())) {
+                        Ok((post_state, gas_used)) => {
                             let Block { mut header, body, .. } = block;
 
                             // clear all transactions from pool
                             pool.remove_transactions(body.iter().map(|tx| tx.hash));
 
-                            header.receipts_root = if res.receipts().is_empty() {
+                            header.receipts_root = if post_state.receipts().is_empty() {
                                 EMPTY_RECEIPTS
                             } else {
-                                let receipts_with_bloom = res
+                                let receipts_with_bloom = post_state
                                     .receipts()
                                     .iter()
                                     .map(|r| r.clone().into())
                                     .collect::<Vec<ReceiptWithBloom>>();
                                 proofs::calculate_receipt_root(receipts_with_bloom.iter())
                             };
-
+                            let transactions = body.clone();
                             let body =
                                 BlockBody { transactions: body, ommers: vec![], withdrawals: None };
                             header.gas_used = gas_used;
@@ -223,11 +229,25 @@ where
                                 }
                             }
 
-                            let header = header.seal_slow();
-                            debug!(target: "consensus::auto", header=?header.hash(), "sending block notification");
+                            // seal the block
+                            let block = Block {
+                                header,
+                                body: transactions,
+                                ommers: vec![],
+                                withdrawals: None,
+                            };
+                            let sealed_block = block.seal_slow();
+                            let sealed_block_with_senders =
+                                SealedBlockWithSenders::new(sealed_block, senders)
+                                    .expect("senders are valid");
+                            debug!(target: "consensus::auto", header=?sealed_block_with_senders.hash(), "sending block notification");
+
+                            let chain =
+                                Arc::new(Chain::new(vec![(sealed_block_with_senders, post_state)]));
 
                             // send block notification
-                            let _ = canon_state_notification.send(Arc::new(header));
+                            let _ = canon_state_notification
+                                .send(reth_provider::CanonStateNotification::Commit { new: chain });
                         }
                         Err(err) => {
                             warn!(target: "consensus::auto", ?err, "failed to execute block")
