+++ crates/storage/provider/src/test_utils/events.rs
@@ -1,7 +1,7 @@
 use std::sync::{Arc, Mutex};
 use tokio::sync::broadcast::{self, Sender};
 
-use crate::{CanonStateNotification, CanonStateNotifications, CanonStateSubscriptions, SubChain};
+use crate::{CanonStateNotification, CanonStateNotifications, CanonStateSubscriptions, Chain};
 
 /// A test ChainEventSubscriptions
 #[derive(Clone, Default)]
@@ -12,14 +12,14 @@ pub struct TestCanonStateSubscriptions {
 impl TestCanonStateSubscriptions {
     /// Adds new block commit to the queue that can be consumed with
     /// [`TestChainEventSubscriptions::subscribe_new_blocks`]
-    pub fn add_next_commit(&mut self, new: Arc<dyn SubChain>) {
-        let event = CanonStateNotification::Commit { new: new.clone() };
+    pub fn add_next_commit(&mut self, new: Arc<Chain>) {
+        let event = CanonStateNotification::Commit { new };
         self.canon_notif_tx.lock().as_mut().unwrap().retain(|tx| tx.send(event.clone()).is_ok())
     }
 
     /// Adds reorg to the queue that can be consumed with
     /// [`TestChainEventSubscriptions::subscribe_new_blocks`]
-    pub fn add_next_reorg(&mut self, old: Arc<dyn SubChain>, new: Arc<dyn SubChain>) {
+    pub fn add_next_reorg(&mut self, old: Arc<Chain>, new: Arc<Chain>) {
         let event = CanonStateNotification::Reorg { old, new };
         self.canon_notif_tx.lock().as_mut().unwrap().retain(|tx| tx.send(event.clone()).is_ok())
     }
