+++ crates/storage/provider/src/traits/mod.rs
@@ -38,6 +38,6 @@ pub use executor::{BlockExecutor, ExecutorFactory};
 
 mod chain;
 pub use chain::{
-    CanonStateNotification, CanonStateNotificationSender, CanonStateNotificationSink,
-    CanonStateNotifications, CanonStateSubscriptions, SubChain,
+    CanonStateNotification, CanonStateNotificationSender, CanonStateNotifications,
+    CanonStateSubscriptions,
 };
